feat(react)!: Remove deprecated `getNumberOfUrlSegments` method
