+++ packages/core/src/utils-hoist/index.ts
@@ -162,8 +162,7 @@ export {
   parseBaggageHeader,
 } from './baggage';
 
-// eslint-disable-next-line deprecation/deprecation
-export { getNumberOfUrlSegments, getSanitizedUrlString, parseUrl, stripUrlQueryAndFragment } from './url';
+export { getSanitizedUrlString, parseUrl, stripUrlQueryAndFragment } from './url';
 // eslint-disable-next-line deprecation/deprecation
 export { makeFifoCache } from './cache';
 export { eventFromMessage, eventFromUnknownInput, exceptionFromError, parseStackFrames } from './eventbuilder';
