+++ packages/core/src/utils-hoist/url.ts
@@ -49,17 +49,6 @@ export function stripUrlQueryAndFragment(urlPath: string): string {
 }
 
 /**
- * Returns number of URL segments of a passed string URL.
- *
- * @deprecated This function will be removed in the next major version.
- */
-// TODO(v9): Hoist this function into the places where we use it. (as it stands only react router v6 instrumentation)
-export function getNumberOfUrlSegments(url: string): number {
-  // split at '/' or at '\/' to split regex urls correctly
-  return url.split(/\\?\//).filter(s => s.length > 0 && s !== ',').length;
-}
-
-/**
  * Takes a URL object and returns a sanitized string which is safe to use as span name
  * see: https://develop.sentry.dev/sdk/data-handling/#structuring-data
  */
