+++ packages/core/test/utils-hoist/url.test.ts
@@ -1,9 +1,4 @@
-import {
-  getNumberOfUrlSegments,
-  getSanitizedUrlString,
-  parseUrl,
-  stripUrlQueryAndFragment,
-} from '../../src/utils-hoist/url';
+import { getSanitizedUrlString, parseUrl, stripUrlQueryAndFragment } from '../../src/utils-hoist/url';
 
 describe('stripQueryStringAndFragment', () => {
   const urlString = 'http://dogs.are.great:1231/yay/';
@@ -23,18 +18,6 @@ describe('stripQueryStringAndFragment', () => {
   it('strips query string and fragment from url', () => {
     const urlWithQueryStringAndFragment = `${urlString}${queryString}${fragment}`;
     expect(stripUrlQueryAndFragment(urlWithQueryStringAndFragment)).toBe(urlString);
-  });
-});
-
-describe('getNumberOfUrlSegments', () => {
-  test.each([
-    ['regular path', '/projects/123/views/234', 4],
-    ['single param parameterized path', '/users/:id/details', 3],
-    ['multi param parameterized path', '/stores/:storeId/products/:productId', 4],
-    ['regex path', String(/\/api\/post[0-9]/), 2],
-  ])('%s', (_: string, input, output) => {
-    // eslint-disable-next-line deprecation/deprecation
-    expect(getNumberOfUrlSegments(input)).toEqual(output);
   });
 });
